Refactoring.

+++ programming-c++/datastructures/stack/abstract-stack/array_stack.h
@@ -20,7 +20,7 @@ class ArrayStack : public Stack
 
     private:
         size_t _size;
-        size_t _top;
+        int    _top;
         int   *_buffer;
 };
